requirements.txt
fastapi==0.112.0
sqlalchemy==2.0.31
pydantic==2.8.2
pymysql==1.1.1
uvicorn==0.30.4
passlib[bcrypt]==1.7.4
python-jose==3.3.0
typing-extensions==4.12.2
python-multipart==0.0.9
bcrypt==4.2.0